Restore full refactor changes from stash
