+++ src/features/locations/pages/LocationDetails.jsx
