+++ src/shared/components/WithOfflineGuard.jsx
