+++ src/shared/layout/Layout.jsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
-import Header from "../Header/Header";
-import Footer from "../Footer/Footer";
+import Header from "./Header";
+import Footer from "./Footer";
 
 function Layout() {
   return (
